--- lib/metadata/shared-metadata.ts
@@ -1,11 +1,0 @@
-export const BASE_OG = {
-  locale: "ko_KR",
-  type: "website",
-};
-
-export const METADATA = {
-  title: "OMIN's Blog",
-  description: "주로 학습한 내용 혹은 회고글을 올립니다.",
-  author: "Minseok OH",
-  url: process.env.BASE_URL || "https://omin.dev",
-};
